fixed bug with marking complete, added a confirm message for deletion, assignment complete

+++ public/main.js
@@ -4,7 +4,7 @@ let listContainer = document.getElementById("listContainer");
 function addItem () {
     let idValue = document.getElementById("idValue").value;
     let todoValue = document.getElementById("todoValue").value;
-    axios.post('/api/todos', {id: idValue, todo: todoValue}).then(function(response){
+    axios.post('/api/todos', {id: idValue, todo: todoValue, completed: "not"}).then(function(response){
         listContainer.innerHTML = renderList(response.data);
     });
 }
@@ -12,59 +12,39 @@ function addItem () {
 function editItem () {
     let idValue = document.getElementById("idValue").value;
     let todoValue = document.getElementById("todoValue").value;
-    axios.put(`/api/todos/${idValue}`, {id: idValue, todo: todoValue}).then(function(response){
+    axios.put(`/api/todos/${idValue}`, {id: idValue, todo: todoValue, completed: "not"}).then(function(response){
         listContainer.innerHTML = renderList(response.data);
     });
-
 }
 
 function deleteItem () {
     let idValue = document.getElementById("idValue").value;
     let todoValue = document.getElementById("todoValue").value;
-    axios.delete(`/api/todos/${idValue}`, {id: idValue, todo: todoValue}).then(function(response){
+    var confirmed = confirm("Are you sure you want to delete?");
+    if (confirmed == true) {
+        axios.delete(`/api/todos/${idValue}`, {id: idValue, todo: todoValue}).then(function(response){
+            listContainer.innerHTML = renderList(response.data);
+        });
+    }
+}
+
+
+function markDone () {
+    let idValue = document.getElementById("idValue").value;
+    let todoValue = document.getElementById("todoValue").value;
+    axios.put(`/api/todos/complete/${idValue}`, {id: idValue, todo: todoValue, completed: "complete"}).then(function(response){
         listContainer.innerHTML = renderList(response.data);
     });
 }
 
-
 function renderList (response) {
     let listHTML = response.map(function(item) {
         return `
-        <div class="d-flex justify-content-between m-auto">
-            <p>${item.id} ${item.todo}</p>
-            <a onclick="markDone(event)" class="btn btn-warning  mb-2">Mark Completed</a>
+        <div class="text-center">
+            <p class="${item.completed} m-3">${item.id} ${item.todo}</p>
         </div> 
         `
     });
     return listHTML.join("");
 }
 
-
-
-function markDone() {
-    let idValue = document.getElementById("idValue").value;
-    axios.get('/api/todos').then(function(response){
-        listContainer.innerHTML = renderMarkDone(idValue, response.data);
-    });
-}
-
-function renderMarkDone(idValue, response) {
-    let listHTML = response.map(function(item) {
-        if (item.id === idValue) {
-            return `
-                <div class="d-flex justify-content-between m-auto">
-                    <p style="text-decoration: line-through;">${item.id} ${item.todo}</p>
-                    <a onclick="markDone(event)" class="btn btn-warning  mb-2">Mark Completed</a>
-                </div> 
-                `
-        } else {
-            return `
-                <div class="d-flex justify-content-between m-auto">
-                    <p>${item.id} ${item.todo}</p>
-                    <a onclick="markDone(event)" class="btn btn-warning  mb-2">Mark Completed</a>
-                </div> 
-                `
-        }
-    });
-    return listHTML.join("");
-}
